NX-164: rework uber classloader
