--- components/nexus-core/src/main/java/org/sonatype/nexus/util/NexusUberClassLoaderProvider.java
@@ -1,49 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2007-2013 Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-
-package org.sonatype.nexus.util;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Provider;
-import javax.inject.Singleton;
-
-import org.codehaus.plexus.PlexusContainer;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-/**
- * Provides a ClassLoader which exposes all classes in Nexus core and in installed plugins.
- *
- * @since 2.6
- */
-@Named("nexus-uber")
-@Singleton
-public class NexusUberClassLoaderProvider
-    implements Provider<ClassLoader>
-{
-  private final WholeWorldClassloader classloader;
-
-  // FIXME: ATM we have to use plexus-specific API here to access the ClassWorld
-
-  @Inject
-  public NexusUberClassLoaderProvider(final PlexusContainer container) {
-    checkNotNull(container);
-    this.classloader = new WholeWorldClassloader(container.getContainerRealm().getWorld());
-  }
-
-  @Override
-  public ClassLoader get() {
-    return classloader;
-  }
-}
